Beta 3.1 support.
Adding Beta 3.1 support. 
Thoughts and insights:
Knowing that invoke will release code breaking changes without warning, we should re-evaluate how we interact with this project.
Will continual changes hinder the requirements for developing on this platform?
How can we get better communication as we JUST added this to the community nodes only for a few days later for it to be broken as 3.1 was released.
How can we time/communicate changes to contributors so as to avoid these issues going forward?